% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.R
\name{disedare_pkg1}
\alias{disedare_pkg1}
\title{This function is reading  in the data set.}
\usage{
disedare_pkg1(file_path)
}
\arguments{
\item{filepath}{allows the data to be installed.}
}
\value{
Returning the data set as disedare_data read into R.
}
\description{
This function is reading  in the data set.
}
\examples{
read_excel("Data/Molmol_database_20180716Dina.xlsx")

}
