% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_3.R
\name{disedare_pkg3}
\alias{disedare_pkg3}
\title{My function is making a histogram of the mass_of_bat data.}
\usage{
disedare_pkg3(file_path)
}
\arguments{
\item{mass_of_bat}{is the bats mass without the mass of the bag, as a numeric and without 'NAs'.}
}
\value{
Returning the histogram of the mass_of_bat data.
}
\description{
My function is making a histogram of the mass_of_bat data.
}
\examples{
histogram <- hist(disedare$mass_of_bat)

}
